+++ app/src/main/java/fr/uge/plutus/data/dao/TagDao.kt
@@ -15,7 +15,7 @@ interface TagDao {
     fun retrieveAll(type: String): Flow<List<Tag>>
 
     @Query("SELECT * FROM `Tag` WHERE tagId=:id")
-    suspend fun retrieveById(id: Int): Tag?
+    suspend fun retrieveById(id: Long): Tag?
 
     @Insert
     suspend fun create(tag: Tag): Long
