+++ app/src/main/java/fr/uge/plutus/data/repository/TagRepository.kt
@@ -15,7 +15,7 @@ class TagRepository @Inject constructor(
 
     override suspend fun createTag(tag: Tag): Long = tagDao.create(tag)
 
-    override suspend fun retrieveTag(id: Int) = tagDao.retrieveById(id)
+    override suspend fun retrieveTag(id: Long) = tagDao.retrieveById(id)
 
     override suspend fun updateTag(tag: Tag) = tagDao.update(tag)
